+++ packages/rocketh-deploy/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-deploy
 
+## 0.6.9
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.6.7
+
 ## 0.6.8
 
 ### Patch Changes