+++ packages/rocketh-doc/CHANGELOG.md
@@ -1,5 +1,13 @@
 # rocketh-doc
 
+## 0.6.3
+
+### Patch Changes
+
+- fix types
+- Updated dependencies
+  - rocketh@0.6.7
+
 ## 0.6.2
 
 ### Patch Changes