+++ packages/rocketh-verifier/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-verifier
 
+## 0.6.7
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh@0.6.7
+
 ## 0.6.6
 
 ### Patch Changes